requirements.txt
fastapi
gliner
groq
hf_xet
huggingface_hub
langchain-core
langchain-text-splitters
langgraph
mistralai
neo4j
python-dotenv
python-multipart
sentence-transformers
spacy
spacy-layout
torch
uvicorn